Moving providers into user library
Instead of having the OAuth implementation live inside of the token library, this moves them into a user library for the interfaces that can then be implemented into the separate implementations.

The initial implementation here has the Google and GitHub implementations.
