+++ user-github/README.md
@@ -1,0 +1,3 @@
+# user-github
+
+This library provides the ability for users to log in with GitHub.
