+++ user-google/README.md
@@ -1,0 +1,3 @@
+# user-google
+
+This library provides the ability for users to log in with Google.
