+++ user/src/main/java/com/unitvectory/serviceauthcentral/user/UserContext.java
@@ -11,7 +11,7 @@
  * or implied. See the License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.unitvectory.serviceauthcentral.server.token.model;
+package com.unitvectory.serviceauthcentral.user;
 
 import lombok.Builder;
 import lombok.Value;
@@ -25,27 +25,27 @@ import lombok.Value;
 @Builder
 public class UserContext {
 
-	/**
-	 * The name of the provider for the user.
-	 */
-	private String provider;
+    /**
+     * The name of the provider for the user.
+     */
+    private String provider;
 
-	/**
-	 * The globally unique immutable identifier for the user.
-	 */
-	private String userId;
+    /**
+     * The globally unique immutable identifier for the user.
+     */
+    private String userId;
 
-	/**
-	 * The display name for the user
-	 */
-	private String userName;
+    /**
+     * The display name for the user
+     */
+    private String userName;
 
-	/**
-	 * Gets the userClientId
-	 * 
-	 * @return the userClientId
-	 */
-	public String getUserClientId() {
-		return String.format("user:%s:%s", this.provider, this.userId);
-	}
+    /**
+     * Gets the userClientId
+     * 
+     * @return the userClientId
+     */
+    public String getUserClientId() {
+        return String.format("user:%s:%s", this.provider, this.userId);
+    }
 }
